+++ compression/archive/zip/const.go
@@ -18,7 +18,9 @@ var (
 	// Maximumn archive size
 	defaultMaxArchiveSize = uint64(1 << 30) // 1GB
 	// Maximum supported file size for archive creation
-	defaultMaxFileSize = uint64(250 * 1 << 20) // 250MB
+	defaultMaxFileSize = uint64(250 << 20) // 250MB
 	// Maximum entry count
 	defaultMaxEntryCount = uint64(10000)
+	// Maximum symlink recursion
+	defaultMaxSymlinkRecursion = uint64(5)
 )